reorganized ucare_cli files structure
